fix: restore kit layout (app files under apps/) and apply response-key parser patch
