--- kits/trading-journal-coach/app/globals.css
@@ -1,92 +1,0 @@
-:root {
-  --bg: #0b0e14;
-  --panel: #141922;
-  --panel-2: #1b212c;
-  --border: #262e3a;
-  --text: #e6e9ef;
-  --muted: #8b93a3;
-  --accent: #4f8cff;
-  --green: #2ecc8f;
-  --red: #ff5c6c;
-  --amber: #ffb547;
-  --radius: 12px;
-}
-
-* { box-sizing: border-box; }
-
-html, body {
-  margin: 0;
-  padding: 0;
-  background: var(--bg);
-  color: var(--text);
-  font-family: ui-sans-serif, system-ui, -apple-system, "Segoe UI", Roboto, Helvetica, Arial, sans-serif;
-  -webkit-font-smoothing: antialiased;
-}
-
-.wrap { max-width: 1080px; margin: 0 auto; padding: 28px 20px 80px; }
-
-.header { display: flex; align-items: center; justify-content: space-between; margin-bottom: 8px; gap: 12px; }
-.brand { display: flex; align-items: center; gap: 10px; }
-.brand h1 { font-size: 20px; margin: 0; letter-spacing: -0.02em; }
-.dot { width: 10px; height: 10px; border-radius: 50%; background: var(--accent); box-shadow: 0 0 12px var(--accent); }
-.sub { color: var(--muted); font-size: 14px; margin: 2px 0 24px; }
-
-.btn { background: var(--accent); color: #fff; border: 0; border-radius: 10px; padding: 10px 16px; font-size: 14px; font-weight: 600; cursor: pointer; transition: filter .15s; }
-.btn:hover { filter: brightness(1.08); }
-.btn:disabled { opacity: .55; cursor: not-allowed; }
-.btn.ghost { background: transparent; border: 1px solid var(--border); color: var(--text); }
-
-.banner { border: 1px solid var(--border); background: var(--panel-2); border-radius: 10px; padding: 10px 14px; font-size: 13px; color: var(--muted); margin-bottom: 18px; }
-.banner b { color: var(--text); }
-.error { border-color: #4a2530; background: #24151a; color: #ffb1ba; }
-
-/* Dropzone */
-.drop { border: 2px dashed var(--border); border-radius: var(--radius); background: var(--panel); padding: 56px 24px; text-align: center; transition: border-color .15s, background .15s; }
-.drop.drag { border-color: var(--accent); background: #101725; }
-.drop h2 { margin: 0 0 6px; font-size: 18px; }
-.drop p { color: var(--muted); margin: 0 0 18px; font-size: 14px; }
-.drop .row { display: flex; gap: 10px; justify-content: center; flex-wrap: wrap; }
-.hint { margin-top: 18px; font-size: 12px; color: var(--muted); }
-.mono { font-family: ui-monospace, SFMono-Regular, Menlo, monospace; font-size: 12px; }
-
-/* Cards grid */
-.cards { display: grid; grid-template-columns: repeat(4, 1fr); gap: 12px; margin: 18px 0; }
-.card { background: var(--panel); border: 1px solid var(--border); border-radius: var(--radius); padding: 14px 16px; }
-.card .label { color: var(--muted); font-size: 12px; text-transform: uppercase; letter-spacing: .04em; }
-.card .value { font-size: 22px; font-weight: 700; margin-top: 6px; letter-spacing: -0.02em; }
-.pos { color: var(--green); }
-.neg { color: var(--red); }
-@media (max-width: 820px) { .cards { grid-template-columns: repeat(2, 1fr); } }
-
-.grid2 { display: grid; grid-template-columns: 1.35fr 1fr; gap: 16px; }
-@media (max-width: 880px) { .grid2 { grid-template-columns: 1fr; } }
-
-.panel { background: var(--panel); border: 1px solid var(--border); border-radius: var(--radius); padding: 18px; }
-.panel h3 { margin: 0 0 12px; font-size: 15px; letter-spacing: -0.01em; }
-
-/* Findings */
-.finding { border: 1px solid var(--border); border-radius: 10px; padding: 12px 14px; margin-bottom: 10px; background: var(--panel-2); }
-.finding .top { display: flex; align-items: center; gap: 8px; margin-bottom: 6px; }
-.finding .title { font-weight: 650; font-size: 14px; }
-.badge { font-size: 11px; font-weight: 700; text-transform: uppercase; padding: 2px 8px; border-radius: 999px; letter-spacing: .03em; }
-.badge.high { background: #3a1b20; color: var(--red); }
-.badge.medium { background: #3a2c14; color: var(--amber); }
-.badge.low { background: #16302a; color: var(--green); }
-.finding .ev { color: var(--muted); font-size: 13px; margin: 4px 0; }
-.rule { font-size: 13px; margin-top: 6px; padding-left: 10px; border-left: 2px solid var(--accent); }
-.score { font-size: 40px; font-weight: 800; letter-spacing: -0.03em; }
-.score-rationale { margin-top: -4px; margin-bottom: 12px; }
-.encourage { margin-top: 12px; font-size: 13px; color: var(--green); border-top: 1px solid var(--border); padding-top: 12px; }
-
-/* Chat */
-.chat { margin-top: 16px; }
-.messages { display: flex; flex-direction: column; gap: 8px; min-height: 60px; margin-bottom: 12px; }
-.msg { max-width: 82%; padding: 9px 13px; border-radius: 12px; font-size: 13.5px; line-height: 1.45; white-space: pre-wrap; }
-.msg.user { align-self: flex-end; background: var(--accent); color: #fff; border-bottom-right-radius: 4px; }
-.msg.bot { align-self: flex-start; background: var(--panel-2); border: 1px solid var(--border); border-bottom-left-radius: 4px; }
-.chatbar { display: flex; gap: 8px; }
-.chatbar input { flex: 1; background: var(--panel-2); border: 1px solid var(--border); border-radius: 10px; padding: 11px 13px; color: var(--text); font-size: 14px; }
-.chatbar input:focus { outline: none; border-color: var(--accent); }
-.suggest { display: flex; gap: 6px; flex-wrap: wrap; margin-bottom: 10px; }
-.chip { font-size: 12px; color: var(--muted); border: 1px solid var(--border); border-radius: 999px; padding: 4px 10px; cursor: pointer; background: transparent; }
-.chip:hover { color: var(--text); border-color: var(--accent); }
