--- kits/trading-journal-coach/next.config.mjs
@@ -1,6 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  reactStrictMode: true,
-};
-
-export default nextConfig;
