--- kits/trading-journal-coach/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "trading-journal-coach",
-  "version": "1.0.0",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "lamatic": "latest",
-    "next": "14.2.5",
-    "papaparse": "^5.4.1",
-    "react": "^18",
-    "react-dom": "^18",
-    "recharts": "^2.12.7"
-  },
-  "devDependencies": {
-    "@types/node": "^20",
-    "@types/papaparse": "^5.3.14",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
-    "eslint": "^8",
-    "eslint-config-next": "14.2.5",
-    "typescript": "^5"
-  }
-}
